Remove unfinished wantarray support.
In the beginning I had the great idea that I would allow you to
perform 'nowait' calls if you didn't include a callback/cv and weren't
waiting for a return value.

In fact, the support for these calls in the protocol is problematic.
You could still get back an error response from a nowait call, which,
if unnoticed, would corrupt your next request, that sort of thing.  So
I'm just going to acknowledge defeat and remove any suggestion of
support for this.