Decouple serializer and compressor.
It was always a bit of an error to mash these two together, but it was
convenient.  Now, in order to support other serializers, *without*
duplicating compression code (and, eventually, to support alternative
compression strategies), we will split them apart.

This adds some test cases for the individual parts, and necessitated a
rename of the test for the overall client, but is still pretty
straightforward.