Use weak references in some places.
This was a bit of a goof to implement, inspired by IO::Async's
_capture_weakself.

Looking at the code it seems like this shouldn't be necessary or
helpful---there are no (obvious) circular references that we don't try
to break by hand---but using a weak ref to $self in several anonymous
subroutines lowers memory use in one of our servers by about 5MB.

Even more strangely, the savings seem to come from the changes in
Memcached::Client, whose uses do not, to my eye, have any potential
for circularity, while Memcached::Client::Connection, which actually
have some circularity ($self->{handle} points to an AnyEvent::Handle,
some of whose callbacks contain references to $self, though we are
good about deleting $self->{handle} when appropriate) shows no benefit
from the changes.

Anyway, it seems to work, it seems to help, it goes in.