The giant request refactor
Refactored the idea of a 'request' out into its own class (and
subclasses), to make the interface between it and the connection and
the protocol a little cleaner and more regular.

This is the first iteration of the refactor that is working---that is,
it passes all tests.  Now to further fasterize it.