Fix test suite to try and accomodate older memcached versions.
I finally realized that the reason I was failing against the CPAN testers
was because those hosts that actually had memcached installed actually
had an old one, and I was making requests it couldn't handle.

So, I've reworked the test suite to, as best I can, eliminate queries
that shouldn't work against a particular release.

Of course, documentation of when stuff was added is kinda spotty, so I
may have missed things, or may have eliminated too much (though I've
tried to be conservative).

Perhaps one day I'll make the client itself adaptive---checking the
version number when it connects and marking certain features as errors
without having to even try them.  But that is down the road.